Reads a single FRE zip file downloaded from Bovespa

Reads a single FRE zip file downloaded from Bovespa

gdfpd.read.fre.zip.file(my.zip.file, folder.to.unzip = tempdir())

Arguments

my.zip.file
Full path to zip file

folder.to.unzip
Folder to unzip files (default = tempdir())

Value

A list with several dataframes containing financial statements

Examples

my.f <- system.file('extdata/FRE_6629_HERC_2010-12-31.zip', package = 'GetDFPData')

my.l <- gdfpd.read.fre.zip.file(my.f)
print(my.l)

#> $df.stockholders
#>     type.register      id.person id.nationality          id.state id.country
#> 1       Acionista 88145669000112     BRASILEIRA Rio Grande do Sul         NA
#> 2       Acionista 86816527000104     BRASILEIRA Rio Grande do Sul         NA
#> 3       Acionista 87431490000169     BRASILEIRA Rio Grande do Sul         NA
#> 4       Acionista 92667526000148     BRASILEIRA Rio Grande do Sul         NA
#> 5       Acionista                    BRASILEIRA              <NA>         NA
#> 6          Outros                          <NA>              <NA>         NA
#> 7 AcoesTesouraria           <NA>           <NA>              <NA>         NA
#> 8           Total           <NA>           <NA>              <NA>         NA
#>                                     name.stockholder type.stockholder
#> 1                          HERCULES S C DE PART LTDA         Juridica
#> 2                                  ZHEPAR PARTC.LTDA         Juridica
#> 3                     ELECE ADM E PARTICIPAÇÕES LTDA         Juridica
#> 4 FUNDAÇÃO FUNCIONÁRIOS EMPRESAS ZIVI E HERCULES S/A         Juridica
#> 5                                             OUTROS         Juridica
#> 6                                             OUTROS         Juridica
#> 7                                               <NA>           Nenhum
#> 8                                               <NA>           Nenhum
#>   qtd.ord.shares perc.ord.shares qtd.pref.shares perc.pref.shares
#> 1        3700344       72.050000               0         0.000000
#> 2             17        0.000000         2795960        27.500000
#> 3         667249       12.990000          868055         8.540000
#> 4         690474       13.440000           25921         0.260000
#> 5          77866        1.520000         6474114        63.700000
#> 6              0        0.000000               0         0.000000
#> 7              0        0.000000               0         0.000000
#> 8        5135950      100.000000        10164050       100.000000
#>   controlling.stockholder
#> 1                    TRUE
#> 2                    TRUE
#> 3                   FALSE
#> 4                   FALSE
#> 5                   FALSE
#> 6                   FALSE
#> 7                   FALSE
#> 8                   FALSE
#> 
#> $df.capital
#>   stock.type stock.class qtd.issued
#> 1         ON           0          0
#> 2         PN           0          0
#> 
#> $df.stock.values
#>   stock.class stock.type max.price min.price avg.price flag.missing.avg.price
#> 1           0         ON       0.0      0.00      0.00                     NA
#> 2           0         PN       0.0      0.00      0.00                     NA
#> 3          21         PN       0.6      0.44      0.52                   TRUE
#>   qtd.issued
#> 1          0
#> 2          0
#> 3         NA
#> 
#> $df.mkt.value
#>   mkt.avg.value mkt.min.value mkt.max.value
#> 1            NA            NA            NA
#> 
#> $df.increase.capital
#> NULL
#> 
#> $df.capital.reduction
#> NULL
#> 
#> $df.compensation
#> NULL
#> 
#> $df.compensation.summary
#>   level.remuneration qtd.members qtd.remunerated.members max.remuneration
#> 1 Management Council           3                      NA                0
#>   mean.remuneration min.remuneration
#> 1                 0                0
#>                                                                                             observations
#> 1 Os Administradores são corporativos e são Remunerados pela Coligada Mundial S/A - Produtos de Consumo.
#> 
#> $df.transactions.related
#> NULL
#> 
#> $df.other.events
#> NULL
#> 
#> $df.stock.repurchases
#> NULL
#> 
#> $df.debt.composition
#>   type.debt type.debt.guarantee debt.value.under.1.year debt.value.1.to.3.years
#> 1        NA                  NA                 2806.41                13449386
#>   debt.value.3.to.5.years debt.value.more.5.years debt.total
#> 1                38989999               249519614  301961805
#> 
#> $df.board.composition
#>                      person.name  person.cpf person.profession
#> 1             Julio Cesar Camara 43837387020          Contador
#> 2 Marcelo de Fagondes de Freitas 52694402020          Contador
#> 3          Cristiano Jacó Renner 22131833091        Engenheiro
#> 4             Paulo Roberto Leke   198676034        Economista
#> 5           Michael Lenn Ceiltin 29599660072        Engenheiro
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               person.cv
#> 1 1983-1984 - Companhia Sudan de Produtos de Tabaco (RS); 1984-1986 – diversas Funções Banco Real S/A; 1986-1991 Diversas Funções Companhia de Cigarros Souza Cruz (RS); 1991-1992 – Supervisor Financeiro Companhia Souza Cruz Industrial e Comercio (RJ); 1992-1996 – Analista de Modelagem Financeira e Relatórios Gerenciais, CECRISA Revestimentos Cerâmicos S/A (SC) e Professor de Análise de Balanços na Faculdade de Criciúma (SC); 1996-1998 – Consultor Empresarial Galeazzi & Associados (SP); 1998-2007 Diretor de Administrativo e Financeiro do Grupo Mundial e atualmente é Diretor de Planejamento e Controle das empresas Mundial S/A e Hercules S/A.
#> 2                                                                                                                                                                                                  1985-1986 – Auxiliar Administrativo na empresa Odorico Monteiro S/A; em 1986 inicio atividades no grupo ZIVI nas funções de Sub-Chefe do Contas a receber de 1986 a 1994, Analista Contábil de 1994-1996, Chefe da Contabilidade Societária de 1996 a 2001, Gerente de Controladoria de 2001 a 2005, Gerente de Controladoria e Finanças de 2005 a 2007, e desde 2007, Eleito Diretor Administrativo e Financeiro das empresas Mundial S/A e Hercules S/A atualmente
#> 3                                   1974-1982 – Sócio Gerente responsável pela gestão administrativa e financeira da MKS – Engenharia de Qualidade Ltda; 1982-1983 – Assessor da Astra Cia. de Adm. Comércio acompanhar o projeto de expansão e reestruturação societária da empresa LEE S/A Ind. De Confecções; 1993 – ingressou nas Lojas Renner S/A e Renner Financiadora S/A; 1984-2004 – Diretor da Renner Financiadora, transformada em 1991 em Banco A J. Renner S/A . 1986-1998 – Diretor de Astra – Cia de Adm e Comércio; 1991-1998 – Diretor Presidente de Lojas Renner S/A; 1999 – Diretor da Textil RV Ltda, atualmente Diretor do Shopping DC Navegantes.
#> 4                                                                                                                                                               : 1977-1982 – Gerente Geral Administrativo Financeiro Grupo Gerdau; 1982-1989 – Diretor Administrativo Financeiro e de Relações com o Mercado na RIOCELL AS; 1990-1994 – Diretor Corporativo Grupo Joaquim Oliveira; 1994-1997 – Diretor Corporativo de Finanças e Controle do Grupo Eberle Mundial; 1997 Sócio de Leke + Gross Consultores Associados; Atividades Atuais: Membro do Conselho de Administração de Lojas Colombo S.A; membro do Conselho de Administração de Mundial S.A e Hercules S.A.
#> 5                                                                                                                                                                                                                         Iniciou suas atividades nas empresas do grupo Zivi que hoje é Grupo Mundial das empresas Mundial S/A e Hercules S/A em 1985, como Gerente dos departamentos de Engenharia Industrial e de Planejamento de Recursos de Manufatura, em 1993 foi Eleito Membro do Conselho de Administração como Vice-Presidente e Diretor Superintendente e atualmente é Presidente do Conselho, Diretor Superintendente e Diretor de Relações com Investidores
#>   person.dob code.type.board    desc.type.board
#> 1         NA               1           Director
#> 2         NA               1           Director
#> 3         NA               2 Management Council
#> 4         NA               2 Management Council
#> 5         NA               3          Counselor
#>                                                              desc.type.board2
#> 1  Diretor de Planejamento e Controle das empresas Mundial S/A e Hercules S/A
#> 2 Diretor Administrativo e Financeiro das empresas Mundial S/A e Hercules S/A
#> 3                                                                        <NA>
#> 4                                                                        <NA>
#> 5                                                     Diretor Superintendente
#>   code.type.job
#> 1            19
#> 2            19
#> 3            22
#> 4            21
#> 5            39
#>                                                              desc.job
#> 1                                                       Não se aplica
#> 2                                                       Não se aplica
#> 3                                                       Não se aplica
#> 4                                                       Não se aplica
#> 5 Presidente do Conselho, Diretor de Relacionamento com Investidores.
#>   date.election date.effective mandate.duration ellected.by.controller
#> 1    2010-07-05     2010-07-05            1 ano                   TRUE
#> 2    2010-07-05     2010-07-05            1 ano                   TRUE
#> 3    2010-07-05     2010-07-05            1 ano                   TRUE
#> 4    2010-07-05     2010-07-05            1 ano                   TRUE
#> 5    2010-07-05     2010-07-05            1 ano                   TRUE
#>   qtd.consecutive.mandates percentage.participation
#> 1                       NA                       NA
#> 2                       NA                       NA
#> 3                       NA                       NA
#> 4                       NA                       NA
#> 5                       NA                       NA
#> 
#> $df.committee.composition
#> NULL
#> 
#> $df.family.relations
#> NULL
#> 
#> $df.family.related.companies
#> NULL
#> 
#> $df.auditing
#>        auditor.name   auditor.cnpj contract.first.date contract.last.date
#> 1 Directa Auditores 11245719000109          2005-05-04               <NA>
#>                            description.contract
#> 1 Auditoria geral das Demonstrações Contábeis. 
#>                                                compensation
#> 1 R$ 23.320,00 (Vinte e três mil, trezentos e vinte reais).
#>   justification.substitution reason.discordance
#> 1                         NA                 NA
#> 
#> $df.responsible.docs
#>   person.cod          person.name                                   person.job
#> 1          4 Michael Lenn Ceitlin Diretor Presidente/Relações com Investidores
#> 
#> $df.stocks.details
#>   type.stock.id type.stock.text tag.along preferential.code
#> 1             1       Ordinária         0                 0
#> 2             2    Preferencial         0                 1
#>       preferential.text
#> 1                  <NA>
#> 2 Preferencial Classe A
#>                                                                                                                            divident.text
#> 1       As ações Ordinárias asseguram a distribuição de um dividendo mínimo obrigatório, correspondendo a 25% do lucro líquido ajustado.
#> 2 Aos acionistas é assegurado,anualmente,distribuição de dividendos mínimos obrigatório, correspondendo a 25% do lucro líquido ajustado.
#>   flag.voting.rights flag.voting.text flag.conversibility other.info.text
#> 1                  1            Pleno                 Não  Não apresenta.
#> 2                  2      Sem Direito                 Não  Não apresenta.
#> 
#> $df.dividends.details
#> # A tibble: 0 x 0
#>